requirements.txt
PyMuPDF>=1.24.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-postgres>=0.0.12
psycopg[binary]>=3.1.0
openai>=1.0.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
truststore
